Fixes Sprite IDs not being assigned correctly on newer versions of Unity.
Updates the project to use the new UnityEditor.U2D.Sprites library when creating the sprite instances for the texture importer.
The previous SpriteMetaData class is now marked as obsolete and was not assigning IDs to sprites causing them to be unable to be selected individually in the Unity editor.
Here is more info about the issue:
https://forum.unity.com/threads/spritemetadata-textureimporter-potential-bug-in-unity-2021-2-16f1.1259632/

And thank you for creating this asset! :smile: